🚚️ api: move controller to src/enrolment context
