+++ api/src/certification/enrolment/application/country-controller.js
@@ -1,5 +1,5 @@
-import { usecases } from '../../domain/usecases/index.js';
-import * as countrySerializer from '../../infrastructure/serializers/jsonapi/country-serializer.js';
+import { usecases } from '../domain/usecases/index.js';
+import * as countrySerializer from '../infrastructure/serializers/country-serializer.js';
 
 const findCountries = async function (_request, _h, dependencies = { countrySerializer }) {
   const countries = await usecases.findCountries();
